Rename google_guest_utils to utils
