+++ google_guest_agent/addresses.go
@@ -25,7 +25,7 @@ import (
 	"strings"
 	"time"
 
-	utils "github.com/GoogleCloudPlatform/guest-agent/google_guest_utils"
+	"github.com/GoogleCloudPlatform/guest-agent/utils"
 	"github.com/GoogleCloudPlatform/guest-logging-go/logger"
 )
