+++ google_guest_agent/non_windows_accounts.go
@@ -26,7 +26,7 @@ import (
 	"strconv"
 	"strings"
 
-	utils "github.com/GoogleCloudPlatform/guest-agent/google_guest_utils"
+	"github.com/GoogleCloudPlatform/guest-agent/utils"
 	"github.com/GoogleCloudPlatform/guest-logging-go/logger"
 )
