+++ google_guest_agent/windows_accounts.go
@@ -28,7 +28,7 @@ import (
 	"math/big"
 	"reflect"
 
-	utils "github.com/GoogleCloudPlatform/guest-agent/google_guest_utils"
+	"github.com/GoogleCloudPlatform/guest-agent/utils"
 	"github.com/GoogleCloudPlatform/guest-logging-go/logger"
 )
